change to float

+++ inc/description.h
@@ -6,7 +6,7 @@
 
 #include <opencv2/opencv.hpp>
 
-std::vector<cv::Mat> get_descriptors(const cv::Mat&, const std::vector<std::tuple<double, double>>&,
-                                     const std::vector<double>&);
+std::vector<cv::Mat> get_descriptors(const cv::Mat&, const std::vector<std::tuple<float, float>>&,
+                                     const std::vector<float>&);
 
 #endif
